requirements.txt
pandas
pytorch_lightning==1.2.10
transformers==4.5.1
torchtext==0.10.1
sentence_transformers